Use cut instead of awk in the install scripts
If the install scripts rely on awk, they malfunction when run on a
device with a limited environment where awk is not available. This patch
replaces the use of awk with cut, which is expected to be more
ubiquitous than awk (and as a bonus also more lightweight on footprint,
execution time and command line length).

Change-Id: I035bf20cb891402677d0a8fbd49ca6fcd099341d
Signed-off-by: Joakim Roubert <joakimr@axis.com>