Let install script(s) fail on unsupported systems
Add check not only if to use Darwin or Linux, but also make sure not to
install the Linux amd64 binaries on e.g. a Linux armv7l system. Then a
proper (?) error message will be displayed and the script aborts.
When/if more systems are supported, they are easily added with this
structure.

Change-Id: Ia0b23785cb89a6620fcbba27aff8270bcf4008ff
Signed-off-by: Joakim Roubert <joakimr@axis.com>