Add install instructions using Homebrew
The existing instructions in the "Getting Started"
section of the documentation cover the install
script and downloading the CLI binary manually.
However for users with Homebrew it may be easier
to install the CLI this way.

This change adds alternative instructions to
install the Linkerd CLI using Homebrew to "Step
1" of the "Getting Started" documentation.

Signed-off-by: James Laverack <james@jameslaverack.com>