Add version label to ease SMI TrafficSplit config
The SMI `TrafficSplit` resouce makes it easy to configure canary releases
in Linkerd. It relies on a `Service` that resolves to all versions
of an application and each version having their own `Service` that
resolves to only endpoints for that version.

Since the current `Deployment` manifests at https://run.linkerd.io/emojivoto.yml
only include the `app` label, you have to redeploy the service you want
to demo `TrafficSplit` with with a `version` label in order to get it
to work properly.

Adding a `version` label will make it so we do not have to redeploy the
current version.

Done in: https://github.com/BuoyantIO/emojivoto/pull/90

Signed-off-by: Noah Krause <krausenoah@gmail.com>