Mute error when attempting to move unexisting file
Followup to #785

When we move the downloaded binary to its final location, we first try
with the file using the architecture suffix, and if that doesn't work we
fallback to not use that suffix. That works, but if the first try fails
we're showing an ugly error:
```
$ curl -sL https://run.linkerd.io/install-edge|sh
...
Download complete!
Validating checksum...
Checksum valid.
mv: cannot stat '/tmp/linkerd2.qY5SfJ/linkerd2-cli-edge-20.7.5-linux-amd64': No such
file or directory
Linkerd edge-20.7.5 was successfully installed 🎉
Add the linkerd CLI to your path with:
  export PATH=$PATH:/home/alpeb/.linkerd2/bin
  Now run:
    linkerd check --pre                     # validate that Linkerd can
    be installed
      linkerd install | kubectl apply -f -    # install the control
      plane into the 'linkerd' namespace
        linkerd check                           # validate everything
        worked!
          linkerd dashboard                       # launch the dashboard
          Looking for more? Visit https://linkerd.io/2/next-steps
```
This PR mutes that error. Tested on Linux and Cygwin.