Rename some anchors and add missing entries to troubleshooting page
When "trust roots" were renamed to "trust anchors" we failed to update the
corresponding anchors in the "troubleshooting" page. I've added some
redirection logic into `linkerd.io/static/checks/index.html` to use the
new anchors while still supporting the old naming.

Also added a new section for the "linkerd-webhooks-and-apisvc-tls"
checks and a missing entry under the HA checks.

This should bring all the linkerd checks in-sync with this doc.