[2.10] Update CLI commands that were moved to Viz
Put `viz` in the CLI commands that were moved to the Viz extension, and add instructions on how to install Viz there where required. Also use Helm configs there where the flags are no longer supported.

I didn't do a thorough follow-up on the instructions for each doc that was updated, so I'm not yet marking them as completed in #5867