feat: blog list add pagination
Had to remove `alias.html` which was overriding the default Hugo alias
template. This was breaking pagination, as for some reason it was set
to set to use `.RelPermalink` instead of the `.Permalink`.

At the time seems it was added because in local development some links were pointing to
linkerd.io instead localhost #71.

Signed-off-by: Christian Mejlak <chris@mejlak.com>