Add new JSON endpoint for extension list
PR#1115 added an `extension-list.yaml` as a way to keep track of all
known linkerd extensions. In issue linkerd/linkerd2#6568, we aim to show
this list of extensions in the linkerd dashboard. Since this list
already resides in the website repo and we'd like the dashboard to get
this list via an API call, it'd be good to serve `extension-list.yaml`
in JSON format.

This change adds a new JSON endpoint with this information. This change
also adds a new field to `extension-list.yaml` that holds a screenshot
file name for each extension. It also adds new png images of these
extension in the website's uploads directory.

Signed-off-by: Dennis Adjei-Baah <dennis@buoyant.io>