fix version linking
Compare paths rather than URLs, so that we can skip links that
have been deprecated rather than generating 404 links.

This is a little brittle because it uses filenames rather than
URLs, so it won't capture things like redirects. But Hugo
doesn't seem to have any function that answers "does this
URL exist in this site?".

Signed-off-by: William Morgan <william@buoyant.io>