Update GitOps guide
I took me some time to figure what `step` was. Later on I checked at it is definitely mentioned in the _Generating your own mTLS root certificates_ guide, but it didn't go top to bottom with the Tasks.

So I figured it would worth mentioning it in this guide as well. Feel free to re-phrase my addition.

Signed-off-by: Bence Ivady <bence.ivady@emarsys.com>