upgrade: remove `$` from bash snippets without output
Markdownlint [suggests][1] that bash snippets showing commands to be run
only include a `$` prefix before the command when the snippet also
includes output from that command. This is why the CI build for the
`alpeb/2.12` branch [is currently failing][2].

Also, there's inconsistent use of `$` in the upgrade notes, in general
--- in some cases, we use `$` in console snippets without output, and in
others, we don't (although markdownlint doesn't catch all of them due to
the use of multi-line commands). I've made the use of `$` more
consistent in the upgrade notes.

[1]: https://github.com/DavidAnson/markdownlint/blob/main/doc/Rules.md#md014---dollar-signs-used-before-commands-without-showing-output
[2]: https://github.com/linkerd/website/runs/7911493490?check_suite_focus=true