tasks: linkify resource names in authz tutorial (#1434)
This branch updates the HTTPRoute authorization tutorial added in #1426
so that the names of policy resources are links to the corresponding
reference docs for those resource types.

This is a separate draft PR because these links will be broken until PR
#1428, which actually adds the reference docs for the policy resource
types, is merged. This PR will either merge into #1426 if #1428 merges
first, or into `shared/docs/2.12` if #1426 merges before #1428.