remove mentions of multi-stage install in 2.12 docs
The multi-stage install feature no longer exists. However, it's
referenced in a few places in the 2.12 documentation, including the
upgrade docs and in the CLI reference (where the sections that are
supposed to contain the multi-stage install subcommands are empty, lol).

This branch removes all those references to multi-stage install.

Fixes #1447