docs: change "Configuring Proxy Concurrency" to reflect reality
Currently, the "Configuring Proxy Concurrency" Tasks documentation
states that the proxy will always use the number of CPU cores available
to the process. This is not actually correct --- if the
`LINKERD2_PROXY_CORES` environment variable is unset, the proxy will
always use a single-threaded runtime. If the environment variable is
set, cgroup CPU limits will _also_ influence the number of worker
threads spawned by the proxy, but when it is not present, the proxy will
always run in single threaded mode.

Whether or not this is the ideal behavior is still open to discussion,
but for now, the documentation should reflect reality. This change
updates the docs to better describe the actual behavior.

Closes linkerd/linkerd2#10698