Subject: Communication Issue Between Viz and External Prometheus
Problem: In the documentation explaining the usage of an external Prometheus, the protocol is missing in provided example. This is causing Viz to be unable to establish communication with the external Prometheus, resulting in a "unsupported protocol" error message in the logs.

Solution: added protocol and basic auth to prometheus url

Fixes #1658

Signed-off-by: JadKHaddad <jadkhaddad@gmail.com>