add top-level HTTPRoute features doc (#1657)
Currently, the website doesn't really have a top-level summary of what
the HTTPRoute resource is and what it can do, similar to the
ServiceProfiles feature page. I thought this would be useful to add, as
it can be used for other documentation to link to, and will also be
useful as a place to collect all the documentation on things users can
do using HTTPRoutes. In addition, we can also add stuff about GAMMA
support and about the difference between `policy.linkerd.io` and
`gateway.networking.kubernetes.io` HTTPRoutes here, as well.

I've also updated some of the other documentation that mentions the
HTTPRoute resource to link to this page.